@@ -7,6 +7,7 @@ import au.com.dius.pact.consumer.junit5.PactTestFor;
 import au.com.dius.pact.core.model.RequestResponsePact;
 import au.com.dius.pact.core.model.annotations.Pact;
 import au.com.dius.pact.core.model.annotations.PactFolder;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.http.HttpResponse;
 import org.apache.http.client.HttpClient;
@@ -14,7 +15,7 @@ import org.apache.http.client.fluent.Executor;
 import org.apache.http.client.methods.HttpPost;
 import org.apache.http.entity.StringEntity;
 import org.apache.http.impl.client.HttpClientBuilder;
-import org.junit.After;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
@@ -24,7 +25,7 @@ import uk.gov.hmcts.reform.et.syaapi.models.ClaimCaseDocument;
 
 import java.io.IOException;
 
-import static org.junit.Assert.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.springframework.http.HttpMethod.POST;
 
 @ExtendWith(PactConsumerTestExt.class)
@@ -39,18 +40,18 @@ class PdfGenerationServiceConsumerTest {
     private static final String SOME_SERVICE_AUTH_TOKEN = "someServiceAuthToken";
 
     @BeforeEach
-    void setUpEachTest() throws InterruptedException, IOException {
+    void setUpEachTest() throws InterruptedException {
         Thread.sleep(2000);
         objectMapper = new ObjectMapper();
     }
 
-    @After
+    @AfterEach
     void teardown() {
         Executor.closeIdleConnections();
     }
 
     @Pact(provider = "etsyaService_pdfgenerationEndpoint", consumer = "et_sya_api_service")
-    RequestResponsePact generatePdfFromTemplate(PactDslWithProvider builder) throws Exception {
+    RequestResponsePact generatePdfFromTemplate(PactDslWithProvider builder) throws JsonProcessingException {
         return builder
             .given("A request to generate a pdf document")
             .uponReceiving("a request to generate a pdf document with a template")
@@ -69,7 +70,7 @@ class PdfGenerationServiceConsumerTest {
 
     @Test
     @PactTestFor(pactMethod = "generatePdfFromTemplate")
-    void verifyGeneratePdfFromTemplatePact(MockServer mockServer) throws Exception {
+    void verifyGeneratePdfFromTemplatePact(MockServer mockServer) throws IOException {
         HttpClient httpClient = HttpClientBuilder.create().build();
 
         HttpPost request = new HttpPost(mockServer.getUrl() + PDF_URL);
@@ -82,10 +83,10 @@ class PdfGenerationServiceConsumerTest {
         String responseContentType = generateDocumentResponse.getEntity().getContentType().toString();
 
         assertEquals(200, generateDocumentResponse.getStatusLine().getStatusCode());
-        assertEquals("Content-Type: application/pdf", responseContentType);
+        assertEquals("Content-type: application/pdf", responseContentType);
     }
 
-    protected String createJsonObject(Object obj) throws Exception {
+    protected String createJsonObject(Object obj) throws JsonProcessingException {
         return objectMapper.writeValueAsString(obj);
     }
 
